+++ src/tasks/helpers/random.ts
@@ -1,0 +1,2 @@
+export const random = (min: number, max: number) =>
+  Math.round((Math.random() * (max - min) + min) * 1000)
