+++ src/tasks/issue-token/sub-tasks/index.ts
@@ -1,0 +1,2 @@
+export * from "./configure-issuer"
+export * from "./create-wallets"
